Seperating bedrock specific utils to new utils file , taken out bedrock model related configs to a new config file which will serve as example, added new section atthe end of readme to address this new integration with guidelines on how to use LLMs hosted on Amazon Bedrock
