+++ utils/completion.py
@@ -13,6 +13,8 @@ import boto3
 from glob import glob
 from tqdm import tqdm
 
+from utils.bedrock_utils import create_llama3_body, create_nova_messages, extract_answer
+
 # API setting constants
 API_MAX_RETRY = 3
 API_RETRY_SLEEP = 10
@@ -1077,38 +1079,6 @@ def chat_completion_aws_bedrock_llama(messages, api_dict=None, aws_region="us-we
 
     return output
 
-def create_llama3_body(messages, max_gen_len=2048, temperature=0.0, top_p=0.9, top_k=1):
-    """
-    Create a request body for Llama 3 models.
-
-    Args:
-    messages (list): List of message dictionaries.
-    max_gen_len (int): Maximum generation length.
-    temperature (float): Temperature for sampling.
-    top_p (float): Top-p sampling parameter.
-    top_k (int): Top-k sampling parameter.
-
-    Returns:
-    str: JSON-encoded string representing the request body for Llama 3 models.
-
-    This function formats the input messages into a prompt suitable for Llama 3 models,
-    including specific formatting tags, and creates a structured request body.
-    """
-    prompt = "\n".join([content for message in messages for content in message["content"]])
-    formatted_prompt = f"""
-    <|begin_of_text|>
-    <|start_header_id|>user<|end_header_id|>
-    {prompt.strip()}
-    <|eot_id|>
-    <|start_header_id|>assistant<|end_header_id|>
-    """
-    return json.dumps({
-        "prompt": formatted_prompt,
-        "max_gen_len": max_gen_len,
-        "temperature": temperature,
-        "top_p": top_p,
-    })
-
 @register_api("aws_nova")
 def chat_completion_aws_bedrock_nova( messages, api_dict=None, aws_region="us-west-2", **kwargs):
     """
@@ -1148,8 +1118,6 @@ def chat_completion_aws_bedrock_nova( messages, api_dict=None, aws_region="us-we
     max_tokens= kwargs["max_tokens"]
     model = kwargs["model_id"]
 
-    print(f"From nova chat completion in completion.py {temperature}, '  ', {max_tokens}, '  ', {model} \n")
-
     # Retry logic for API calls
     for _ in range(API_MAX_RETRY):
         try:
@@ -1183,67 +1151,6 @@ def chat_completion_aws_bedrock_nova( messages, api_dict=None, aws_region="us-we
             time.sleep(API_RETRY_SLEEP)
 
     return output
-
-def extract_innermost_text(content):
-    """Recursively extract the innermost `text` value from a deeply nested structure."""
-    if isinstance(content, list) and content:  # If content is a list, dive into the first element
-        return extract_innermost_text(content[0])
-    elif isinstance(content, dict) and "text" in content:  # If content is a dictionary, dive into the 'text' key
-        return extract_innermost_text(content["text"])
-    elif isinstance(content, str):  # Base case: return the string when reached
-        return content
-    return ""  # Fallback in case the structure is invalid
-
-def create_nova_messages(messages):
-    """
-    Create messages array for Nova models from conversation
-
-    Args:
-    conv (object): Conversation object containing messages
-
-    Returns:
-    list: List of formatted messages for Nova model
-    """
-    messages_formatted = []    
-    # Format the first message with template
-    for mesg in messages:        
-        # Transform the message
-        transformed_message = {
-            "role": mesg["role"],
-            "content": extract_innermost_text(mesg["content"])
-        }
-        messages_formatted.append({
-            "role": "user",
-            "content": [
-                { 
-                    "text": transformed_message['content']
-                }
-            ]
-        })    
-    return messages_formatted
-
-import sys
-import re
-
-def extract_answer(text):
-    """
-    Extract the content after the </think> tag.
-    
-    Args:
-        text (str): Input text that may contain a </think> tag
-        
-    Returns:
-        str: Text after the </think> tag, or the original text if tag not found
-    """
-    # Look for the </think> tag
-    match = re.search(r'</think>(.*)', text, re.DOTALL)
-    
-    if match:
-        # Return only the content after the tag
-        return match.group(1).strip()
-    else:
-        # If tag not found, return empty string
-        return ""
 
 @register_api("aws_deepseek")
 def chat_completion_aws_bedrock_deepseek(messages, api_dict=None, aws_region="us-west-2", **kwargs):
